requirements.txt
annotated-types==0.7.0
anyio==4.11.0
bcrypt==4.1.3
certifi==2025.10.5
click==8.3.0
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.118.0
fastapi-cli==0.0.13
fastapi-cloud-cli==0.3.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
orjson==3.11.3
packaging==25.0
passlib==1.7.4
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycryptodome==3.23.0
pydantic==2.12.0
pydantic-extra-types==2.10.5
pydantic-settings==2.11.0
pydantic_core==2.41.1
Pygments==2.19.2
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
PyYAML==6.0.3
rich==14.1.0
rich-toolkit==0.15.1
rignore==0.7.0
rsa==4.9.1
sentry-sdk==2.40.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.48.0
typer==0.19.2
typing-inspection==0.4.2
typing_extensions==4.15.0
ujson==5.11.0
urllib3==2.5.0
uvicorn==0.37.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
